Avoid dropping requests when we reboot main.py
Remove the skip_existing parameter to have PRAW retrieve the 100 most
recent replies every time it starts up. Skip over the ones we've
already processed (marked as read). This makes sure we don't drop
requests when we bring down the bot, upgrade it, and bring it back up
within a few minutes.