HUGE LIMIT POSTING UPDATE
Okay, this is quite huge, lemme break it down.
Now, submitter is connected to the database because we filter all our new submissions. So. if you want to post a meme, you would need a minimum of 1000 MemeCoins to do that or 10%. If your balance is exactly 1000, the bot will take all of it. If you have 2000, it will take only 1000. But if your balance is greater than 10,000 you will start paying full 10% fee. If you have 999,999 MemeCoins, you will pay 99,999. This is the huge update. Docker dependency added successfully. Also, comments with balances are bold now. This needs a review.

Tests can be found on the [test subreddit](https://reddit.com/r/memeinvestor_test)