user: Render time of investment using locale date format
The primary impetus here is that invTime.getMinutes() does not return
a zero-prefixed two digit number for investments in the first 10
minutes of the hour. Seeing as most browsers know how to render based
on locale string, this seems like a very logical change, as it means
that this function no longer needs to deal with its own time style.

(In lieu of another library e.g. moment, this is probably the easiest
solution.)

Signed-off-by: Kristofer Rye <kristofer.rye@gmail.com>