Added a test variable.
This variable is needed for testing purposes. To add flairs, I had modify comment_worker.py directly. Our testing fails because comment_worker tries to initialize praw.Reddit instance that should not be created. To avoid that, I threw some if statements to check if the bot is in a testing environment.