Added new columns to the Firm table.
firm.size -> shows the size of the firm (all members)
firm.execs -> number of executives in the firm (CEO not included)
firm.boost -> the profit boost that firm members git from their investments (1 is 0%, 1.1 is 10%), meant to be an upgrade
firm.tax -> the tax that all members of a firm should pay from their profits, needs to be regulated and limited, not 0% and not 100%, add validation steps in the code!
firm.rank -> firm's rank and the upgrade step, 1st lvl, 2nd lvl, and etc.