Added a dummy service and script for logging.
Sometimes we need to execute manual scripts. To do that we always used one of the python-based containers. If none of the services are running, it is becoming inconvenient at executing the scripts in python. This update add a dummy containers that will be in a network with the database and will be running tester.py script to stay alive.