Added new caddy config file. Goaccess is used in real-time (hopefully)
Alright, with this update I want to have more control over the visitors or our website.

Caddy is our webserver and its logs can be used with goaccess to get a nice visualization.

I added a new subdomain goaccess.meme.market to get the current state.

But I still did not figure out how to make it real time. Because the goaccess page doesn't update on every webbrowser visit. It only updates when we restart the container.